pyproject.toml
[tool.poetry]
name = "integracao-rd"
version = "0.1.0"
description = ""
authors = ["GabMeula07 <gabrielqueirozmeula@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pandas = "^2.2.2"
sqlalchemy = "^2.0.31"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
